backend/requirements.txt
aioredis==2.0.1
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
billiard==4.2.1
boto3==1.38.42
botocore==1.38.46
cachetools==6.2.4
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.3
Deprecated==1.2.18
deprecation==2.1.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fsspec==2025.12.0
greenlet==3.2.3
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
kombu==5.5.4
limits==5.4.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
postgrest==2.27.0
prompt_toolkit==3.0.51
propcache==0.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.9.1
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.2.5
pyroaring==1.0.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
realtime==2.27.0
redis==6.2.0
requests==2.32.4
resend==2.19.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.46.2
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
stripe==12.4.0
style==1.1.0
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
update==0.0.1
urllib3==2.5.0
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.22.0